Add Global.java on project
Global.java it's a file responsible to say what happen when server
start operating, and when server stops operating. This can be useful to
load default configuration Objects on database, or initializing some
Singleton instances.

For now, this file only uses Logger to inform when server loaded and
when stopped.

 - Add Global.java using Logger to inform start and end of server