Add Favicons
The favicons are icons displayed on tabs, indicating visually which
page it's loaded. These icons were got on Internet, in a license that
requires a credit to author in some place of the page.

The place for the credit will be at the bottom of the page, in
preference.

Also, I've added in 3 dimensions, for more purposes.

 - Added favicon in 3 different sizes