Add Navbar text content
This JSON will contain all text loaded on the navigation bar on Home
page. It's on this JSON, to easier access on the information, and with
this, it's possible to alter the text, without compiling project again.

Text here, can be changed later, but for now, it maintain sections
working. Contains the button for logo, with the link to initial section
, contains the button for about section, button for services section,
and button for contact section.

 - Add JSON containing text for navigation bar on home