Add header text content for homepage
On initial section, a header poster will atract more clients for the
website. This poster will contain big title, with a effect phrase,
motivational text, and links to act on Website. Also, the images are
very important.

The images initially are loaded from a picasa album of mine. This can
be changed in other commits.

 - Added header text for homepage initial section