Add banners text contents
All of banners shows a big text with a beautiful image behind. The main
function of these are to call attention, and to differentiate from rest
of page.

There are the banners used as title for every section. But the recall
banner, serve to recall client on the opportunity to use the website.

 - Add banners text content, for all sections