Add Feed views to project
I've added the Feed part of our website, where the main functions will
be available. Until now, it only shows the ads onto the initial section
of the page. All the necessary items were already imported.

The Feed part, will contain a navigation bar, a footer bar, and a part
where all the rest of the site will be.

 - Add Feed views