Mod background on Ads Section
I've choosen a better pattern to serve as background for Ads Section.
This background was made by gradient.quasi.ink and it's a tiny little
piece of a picture of this project. Using a 20 x 15 pixels image, the
site creates a extra blur to transform the image into an amazing
background pattern cover.

 - Modify pattern used on Ads Section