Mod layout of feed
I've modified all of the layout to a pretty and better material design
layout. This layout take consideration for the cake business, using a
brown color to ressemble the chocolate. Also, the pattern behind all
content can give a clean impression of our site.

In resume, the layout follows many things of the previous layout. From
the navbar to the footer. But a few things changed: I've added a logo
to the header with logotype; also the top bar will say have title of
the current section.

 - Incorporate Material Design Lite for the feed