Mod FeedApp to allow modals on project
I've modify some things: First, the HTML will have two separate views
for the section. One views will be the main one, and the another will
contain all necessary modals.

Also, this file will call the init for Modals, when HTML finishes all
loading. This actually are bad implemented, because I use only a
timeout function to ensure this. A better approach would be to really
detect all work on angular to be done, and while this, a loading circle
would appear, only disapearing when modal inits.

 - Adds modal compatibility to FeedApp