Mod controllers to load info for new directives
I've modify the controllers to load new information for the new
directive: the toolbar. First, the method to get title of section, and
a call of this in every section. Second, the toolbar information will
be separated into the labels default for the user btns on a JSON, and
the user name that will be load with the API.

 - Modify controllers to load info for tool-bar