Add directives for Register page
I've added some directives to create a Register page. The needed were
a form-title, to display any interesting welcome message; form-tabs
with the buttons indicating steps; layout-one-card to pack all display
into a big card in front of a ribbon and a background; and the 3 panels
that will get info on User.

 - Add new directives